Updated gui to resize properly added search field

+++ src/main/java/com/gvlocke/weatherapp/Current.java
@@ -56,7 +56,7 @@ public class Current {
         return snowfall;
     }
 
-    public double getWeatherCode() {
+    public int getWeatherCode() {
         return weatherCode;
     }
